Add data notebooks
