+++ common/data/src/interfaces/JsonSchema.ts
@@ -1,0 +1,12 @@
+export interface JsonSchema {
+    $id: string,
+    $schema: string,
+    title: string,
+    type: string,
+    properties: {
+        [key: string]: {
+            type: string,
+            description: string
+        }
+    },
+}
