+++ common/data/src/lib/agg.schema.json
@@ -1,0 +1,32 @@
+{
+	"$id": "https://casimir.co/agg.schema.json",
+	"$schema": "http://json-schema.org/draft-07/schema#",
+	"title": "Agg",
+	"type": "object",
+    "properties": {
+        "type": {
+            "type": "string",
+            "description": "The type of aggregate (e.g. wallet, contract, etc.)"
+        },
+        "address": {
+            "type": "string",
+            "description": "The address of the aggregate"
+        },
+        "first_staked_at": {
+            "type": "date",
+            "description": "The first datestring (MM-DD-YYYY) that a wallet staked"
+        },
+        "total_staked_amount": {
+            "type": "string",
+            "description": "The total amount that a wallet has staked"
+        },
+        "total_staked_duration": {
+            "type": "integer",
+            "description": "The total duration that a wallet has staked"
+        },
+        "auto_staking": {
+            "type": "boolean",
+            "description": "The most recent stake reward compounding selection of a wallet"
+        }
+    }
+}
