--- common/schemas/package.json
@@ -1,18 +1,0 @@
-{
-    "name": "@casimir/schemas",
-    "private": true,
-    "main": "src/index.ts",
-    "scripts": {
-      "dev": "npx esno src/index.ts",
-      "build": "esbuild src/index.ts --bundle --minify --sourcemap --platform=node --target=es2020 --outfile=dist/index.js",
-      "test": "echo \"Error: no test specified\" && exit 1"
-    },
-    "devDependencies": {
-      "@types/node": "^17.0.38",
-      "esbuild": "^0.14.42",
-      "esno": "^0.16.3"
-    },
-    "dependencies": {
-        "@aws-cdk/aws-glue-alpha": "^2.33.0-alpha.0"
-    }
-}
