--- common/schemas/src/interfaces/JsonSchema.ts
@@ -1,12 +1,0 @@
-export interface JsonSchema {
-    $id: string,
-    $schema: string,
-    title: string,
-    type: string,
-    properties: {
-        [key: string]: {
-            type: string,
-            description: string
-        }
-    },
-}
