--- common/schemas/src/lib/event.schema.json
@@ -1,36 +1,0 @@
-{
-	"$id": "https://casimir.co/event.schema.json",
-	"$schema": "http://json-schema.org/draft-07/schema#",
-	"title": "Event",
-	"type": "object",
-    "properties": {
-        "type": {
-            "type": "string",
-            "description": "The type of event"
-        },
-        "datestring": {
-            "type": "date",
-            "description": "The datestring (MM-DD-YYYY) of the event"
-        },
-        "address": {
-            "type": "string",
-            "description": "The address that initiated the event"
-        },
-        "staked_candidate": {
-            "type": "string",
-            "description": "The name of the candidate that received the stake action event"
-        },
-        "staked_amount": {
-            "type": "string",
-            "description": "The amount staked or unstaked in the stake action event"
-        },
-        "staked_duration": {
-            "type": "integer",
-            "description": "The duration of the stake action event"
-        },
-        "auto_stake":{
-            "type": "boolean",
-            "description": "The compounding selection of the stake action event"
-        }
-    }
-}
