+++ infrastructure/cdk/lib/etl/etl-stack.ts
@@ -2,7 +2,7 @@ import { Stack, StackProps } from 'aws-cdk-lib'
 import { Construct } from 'constructs'
 import * as s3 from 'aws-cdk-lib/aws-s3'
 import * as glue from '@aws-cdk/aws-glue-alpha'
-import { jsonToGlueSchema, eventSchema, aggSchema } from '@casimir/schemas'
+import { jsonToGlueSchema, eventSchema, aggSchema } from '@casimir/data'
 
 // Get Glue Columns from JSON Schema for each table
 const eventColumns = jsonToGlueSchema(eventSchema)
